Changes in the backend

+++ server/src/controllers/prospects.controller.js
@@ -7,6 +7,19 @@ const resend = new Resend('re_AFCEBU6U_JwRaBJ3vnDm1oHujCaQYRT1G');
 
 export const getProspects = async (req, res) => {
     try {
+        const { userId } = req
+
+        // Verificar que el usuario sea de tipo admin o advisor
+        const [users] = await pool.query('SELECT * FROM advisors WHERE id = ?', [userId]);
+        if (users.length <= 0) {
+            return res.status(400).send({ error: 'Invalid user id' });
+        }
+
+        const user = users[0];
+        if (!['admin', 'advisor'].includes(user.user_type)) {
+            return res.status(403).send({ error: 'Unauthorized' });
+        }
+
         const [rows] = await pool.query('SELECT * FROM prospects');
         res.json(rows);
     } catch (error) {
@@ -17,9 +30,22 @@ export const getProspects = async (req, res) => {
 export const getProspectById = async (req, res) => {
     try {
         const id = req.params.id;
+        const { userId } = req
+
+        // Verificar que el usuario sea de tipo admin o advisor
+        const [users] = await pool.query('SELECT * FROM advisors WHERE id = ?', [userId]);
+        if (users.length <= 0) {
+            return res.status(400).send({ error: 'Invalid user id' });
+        }
+
+        const user = users[0];
+        if (!['admin', 'advisor'].includes(user.user_type)) {
+            return res.status(403).send({ error: 'Unauthorized' });
+        }
+
         const [rows] = await pool.query('SELECT * FROM prospects WHERE id = ?', [id]);
         if (rows.length <= 0) return res.status(404).json({
-            message: 'Prospecto no Encontrado'
+            message: 'Prospect not found'
         });
         res.json(rows[0]);
     } catch (error) {
@@ -64,16 +90,28 @@ export const updateProspect = async (req, res) => {
         const { id } = req.params;
         const { name, lastname, email, phone_number, age, addresses } = req.body;
 
+        const { userId } = req
+
+        // Verificar que el usuario sea de tipo admin o advisor
+        const [users] = await pool.query('SELECT * FROM advisors WHERE id = ?', [userId]);
+        if (users.length <= 0) {
+            return res.status(400).send({ error: 'Invalid user id' });
+        }
+
+        const user = users[0];
+        if (!['admin', 'advisor'].includes(user.user_type)) {
+            return res.status(403).send({ error: 'Unauthorized' });
+        }
+
+        // Actualizar el prospecto
         const [result] = await pool.query('UPDATE prospects SET name = IFNULL(?, name), lastname = IFNULL(?, lastname), email = IFNULL(?, email), phone_number = IFNULL(?, phone_number), age = IFNULL(?, age), addresses = IFNULL(?, addresses) WHERE id = ?', [name, lastname, email, phone_number, age, addresses, id]);
 
-
         if (result.affectedRows === 0) return res.status(404).json({
-            message: 'Prospect no encntrado'
+            message: 'Prospect not found'
         })
 
         const [rows] = await pool.query('SELECT * FROM prospects WHERE id = ?', [id])
         res.json(rows[0]);
-
 
     } catch (error) {
         console.error(error);
@@ -83,8 +121,21 @@ export const updateProspect = async (req, res) => {
 
 export const deleteProspect = async (req, res) => {
     try {
+        const { userId } = req
+
+        // Verificar que el usuario sea de tipo admin o advisor
+        const [users] = await pool.query('SELECT * FROM advisors WHERE id = ?', [userId]);
+        if (users.length <= 0) {
+            return res.status(400).send({ error: 'Invalid user id' });
+        }
+
+        const user = users[0];
+        if (!['admin', 'advisor'].includes(user.user_type)) {
+            return res.status(403).send({ error: 'Unauthorized' });
+        }
+
+        // Eliminar el prospecto
         const [result] = await pool.query('DELETE FROM prospects WHERE id = ?', [req.params.id]);
-        console.log(result);
 
         if (result.affectedRows <= 0) return res.status(404).json({
             message: 'Prospect not found'
